feat(packaging): updated package.json and bumped deps
As of this update we now have propperly configured `package.json` files
all the way down the dependency tree. What does that mean for you as a
developer?

When you depend on `mu-jquery-app` you're actually more likely to use our
bundled dependencies (as expected, we're a toolkit - not a framework) and
thus you can _almost_ look at `mu-jquery-app` as a bundle of toolkit
modules. With this update this intent is expressed in how utilize
`peerDependencies` combined with `bundledDependencies` in package.json.

So what do you as a developer have to change? Nothing. What you get is a
more sane dependency model that ensures your dependencies stay up do date
and consistent.